dataLine and dataLabel enhancements
- Added functions to create ‘g’ containers for lines and labels that
are called in BaseChart.draw(). This was done to permit more explicit
ordering/layering of these elements w.r.t other SVG elements in the
chart. For example, data labels are now layered such that they are
clearly visible on top of data points, axes, and lines!
- Related to above, replaced .append() with .insert() in a few places
to achieve desired layering of SVG elements.
- Removed removeDataLabel() since removeSelection() does the job